backend/requirements.txt
# backend/requirements.txt

Flask==2.1.0
Werkzeug==2.0.3
psycopg2-binary==2.9.6
redis==4.5.1